- implement more of business logic

+++ src/MyCASINO/_shared/cpp/MyCasinoLib/MyCasinoTransaction.cpp
@@ -12,7 +12,7 @@ MyCasinoTransaction::~MyCasinoTransaction()
 	m_pTransactionDetails = NULL;
 }
 
-DOUBLE MyCasinoTransaction::GetId()
+ULONG MyCasinoTransaction::GetId()
 {
 	return m_id;
 }
